requirements.txt
django
torch
torchvision
pillow
numpy
tqdm